functions/requirements.txt
names
google-cloud-pubsub
google-api-core==2.10.2